Add TeamStatistics class
