+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ require "simplecov"
 SimpleCov.start
 
 require_relative 'stat_tracker_spec'
+require_relative 'team_statistics_spec'
 
 RSpec.configure do |config|
   config.formatter = :documentation
